FFmpegWriterで追加の引数を渡せるようにした
